BufferUnderflowException

+++ src/main/java/com/atguigu/nio/NIOByteBufferPutGet.java
@@ -18,16 +18,9 @@ public class NIOByteBufferPutGet {
         buffer.flip();
 
         System.out.println();
-
         System.out.println(buffer.getInt());
         System.out.println(buffer.getLong());
         System.out.println(buffer.getChar());
-        System.out.println(buffer.getShort());
-
-
-
-
-
-
+        System.out.println(buffer.getLong());
     }
 }
